test: DeckCards 테스트 코드 이동
