+++ README.md
@@ -1,9 +1,9 @@
 # kotlin-blackjack
 
 ## step2
-- [x] 초기 Deck 은 52장의 카드로 이루어진다
-- [x] 초기 Deck 의 각 Suit 카드는 13장씩 있어야 한다
-- [x] 초기 Deck 중복 없는 카드로 이루어져야 한다
+- [x] 초기 DeckCards 은 52장의 카드로 이루어진다
+- [x] 초기 DeckCards 의 각 Suit 카드는 13장씩 있어야 한다
+- [x] 초기 DeckCards 는 중복 없는 카드로 이루어져야 한다
 - [x] Deck 에서 draw 하면 Deck 의 카드 수는 1 감소해야 한다
 - [x] Deck 에서 draw 한 카드는 Deck 에서 사라져야 한다
 - [x] Hand 는 초기에 2 장의 카드로 이루어져야 한다
