+++ src/test/kotlin/blackjack/domain/player/PlayerTest.kt
@@ -87,4 +87,43 @@ class PlayerTest : StringSpec({
         }.isInstanceOf(RuntimeException::class.java)
             .hasMessageContaining("Invalid state transition")
     }
+
+    "Hand 의 sum 이 21인 Player 는 blackjack 상태가 된다" {
+        val hand = Hand(
+            HandCards(
+                mutableListOf(
+                    Card(Suit.Spade, Character.Jack),
+                    Card(Suit.Clover, Character.Ace),
+                )
+            )
+        )
+        val player = Player("aaa", hand)
+
+        player.addCard(Card(Suit.Diamond, Character.Jack))
+
+        hand.valueSum() shouldBe 21
+        hand.isBlackjack() shouldBe true
+        hand.isBust() shouldBe false
+        player.state shouldBe PlayerState.Blackjack
+    }
+
+    "Hand 의 sum 이 21을 초과하는 Player 는 bust 상태가 된다" {
+        val hand = Hand(
+            HandCards(
+                mutableListOf(
+                    Card(Suit.Spade, Character.Two),
+                    Card(Suit.Clover, Character.Jack),
+                )
+            )
+        )
+
+        val player = Player("aaa", hand)
+
+        player.addCard(Card(Suit.Diamond, Character.Jack))
+
+        hand.valueSum() shouldBe 22
+        hand.isBlackjack() shouldBe false
+        hand.isBust() shouldBe true
+        player.state shouldBe PlayerState.Bust
+    }
 })
